feat: add new util file for mapwithjsonpath
